swell-generic: send LVN_ITEMCHANGED notifications when (un)selecting all items
This fixes ListView_SetItemState not sending any notifications when (un)selecting
all items. Like the Windows API and swell-cocoa, one notification is sent per item.

Also selecting all no longer selects the last item when the list is in
single-selection mode. This is on-par with the behavior of the Windows API.